cleanup

+++ kitdsp/src/chorus.cpp
@@ -1,4 +1,5 @@
 #include "kitdsp/apps/chorus.h"
+#include "kitdsp/math/interpolate.h"
 
 namespace kitdsp {
 
@@ -21,7 +22,8 @@ float_2 Chorus::Process(float_2 in) {
 
     float lfoMs = (mLfo.Process() * cfg.delayModMs);
     auto toSamples = [&](float offsetMs) {
-        return clamp<float>((cfg.delayBaseMs + offsetMs) * mSampleRate / 1000.0f, 0.0f, mDelayLine.Size());
+        return clamp<float>((cfg.delayBaseMs + offsetMs) * mSampleRate / 1000.0f, 0.0f,
+                            static_cast<float>(mDelayLine.Size()));
     };
 
     using namespace kitdsp::interpolate;
@@ -43,7 +45,7 @@ float_2 Chorus::Process(float_2 in) {
 }
 
 float Chorus::GetMaxDelayMs() const {
-    return mDelayLine.Size() * 1000.0f / mSampleRate;
+    return static_cast<float>(mDelayLine.Size()) * 1000.0f / mSampleRate;
 }
 
 }  // namespace kitdsp
